+++ src/com/intellij/idea/plugin/hybris/toolwindow/RemoteSolrConnectionDialog.kt
@@ -145,7 +145,7 @@ class RemoteSolrConnectionDialog(
         hostIP = hostTextField.text
         port = portTextField.text
         isSsl = sslProtocolCheckBox.isSelected
-        isWsl = isWslCheckBox.isSelected
+        isWsl = isWslCheckBox?.isSelected ?: false
         solrWebroot = webrootTextField.text
         credentials = Credentials(usernameTextField.text, String(passwordTextField.password))
         this